fix: break thumbnails in sidebar to multiple lines
In the sidebar if when I have more than 3 thumbnails, only the first 3
is visible. That makes very hard to view or edit the faces that are
recognized on the photo.

The fix is to break the thumbnails to new line when needed by adding the
flex-wrap class to the container ul element.